eOrder:update id type and Order.java architecture
Description: also move class DishCategory from Category.java to outer
class file.

+++ src/com/androidodc/eorder/datatypes/Config.java
@@ -2,25 +2,25 @@ package com.androidodc.eorder.datatypes;
 
 public class Config {
 
-    private int mId;
-    private int mConfigId;
+    private long mId;
+    private long mConfigId;
     private String mName;
     private String mValue;
     private String mDescription;
 
-    public int getId() {
+    public long getId() {
         return mId;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.mId = id;
     }
 
-    public int getConfigId() {
+    public long getConfigId() {
         return mConfigId;
     }
 
-    public void setConfigId(int configId) {
+    public void setConfigId(long configId) {
         mConfigId = configId;
     }
